Move SideBar component in Items folder
